+++ src/main/java/com/tb/javaecommerce/service/exception/CategoryNotFoundException.java
@@ -3,7 +3,7 @@ package com.tb.javaecommerce.service.exception;
 public class CategoryNotFoundException extends RuntimeException {
   private static final String CATEGORY_NOT_FOUND_MESSAGE = "Category With ID - %s Not Found";
 
-  public CategoryNotFoundException(int id) {
+  public CategoryNotFoundException(Long id) {
     super(String.format(CATEGORY_NOT_FOUND_MESSAGE, id));
   }
 }
